Feat: 스와이프로 다음 퀴즈 네비게이션 구현
