+++ package.json
@@ -16,6 +16,7 @@
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-swipeable": "^7.0.1",
     "rehype-prism-plus": "^1.6.1",
     "rehype-stringify": "^9.0.3",
     "remark": "^14.0.3",
